Move checking if value is in range
